// src/types/ERROR_CODE.ts
enum ERROR_CODE {
  UNSUPPORTED_MEDIA_TYPE = "UNSUPPORTED_MEDIA_TYPE",
  VALIDATION_ERROR = "VALIDATION_ERROR",
  NOT_FOUND = "NOT_FOUND",
  ROUTE_NOT_FOUND = "ROUTE_NOT_FOUND",
  NOT_IMPLEMENTED = "NOT_IMPLEMENTED",
  UNEXPECTED_ERROR = "UNEXPECTED_ERROR",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",
  METHOD_NOT_ALLOWED = "METHOD_NOT_ALLOWED",
}

export default ERROR_CODE